Added ids, revised drive for little endian

+++ RED/Models/Modules/DriveModel.cs
@@ -2,8 +2,8 @@
 {
     internal class DriveModel
     {
-        internal int SpeedLeft;
-        internal int SpeedRight;
+        internal short SpeedLeft;
+        internal short SpeedRight;
 
         internal int SpeedLimit;
         internal bool UseLegacyDataIds;
